Pull out state module
