+++ escrow/examples/schema.rs
@@ -2,8 +2,8 @@ use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 use std::env::current_dir;
 use std::fs::create_dir_all;
 
-use cw_escrow::contract::State;
 use cw_escrow::msg::{HandleMsg, InitMsg, QueryMsg};
+use cw_escrow::state::State;
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
