+++ escrow/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod contract;
 pub mod msg;
+pub mod state;
 
 #[cfg(target_arch = "wasm32")]
 mod wasm {
