+++ escrow/src/state.rs
@@ -1,0 +1,41 @@
+use cosmwasm_std::{CanonicalAddr, Env, Storage};
+use cosmwasm_storage::{singleton, singleton_read, ReadonlySingleton, Singleton};
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+static CONFIG_KEY: &[u8] = b"config";
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct State {
+    pub arbiter: CanonicalAddr,
+    pub recipient: CanonicalAddr,
+    pub source: CanonicalAddr,
+    pub end_height: Option<u64>,
+    pub end_time: Option<u64>,
+}
+
+impl State {
+    pub fn is_expired(&self, env: &Env) -> bool {
+        if let Some(end_height) = self.end_height {
+            if env.block.height > end_height {
+                return true;
+            }
+        }
+
+        if let Some(end_time) = self.end_time {
+            if env.block.time > end_time {
+                return true;
+            }
+        }
+
+        false
+    }
+}
+
+pub fn config<S: Storage>(storage: &mut S) -> Singleton<S, State> {
+    singleton(storage, CONFIG_KEY)
+}
+
+pub fn config_read<S: Storage>(storage: &S) -> ReadonlySingleton<S, State> {
+    singleton_read(storage, CONFIG_KEY)
+}
